requirements.txt
Flask>=0.10.1
braintree>=3.22.0
MarkupSafe>=0.23
mock>=1.3.0
Werkzeug>=0.7
Jinja2>=2.4
itsdangerous>=0.21
requests<3.0,>=0.11.1
pbr>=0.11
six>=1.7
python-dotenv==0.4.0
gunicorn==19.4.5